requirements.txt
streamlit
easyocr
opencv-python
numpy
pillow
scikit-learn
pymupdf
google-generativeai
python-dotenv